+++ apps/frontend/my-app/src/Component/HumbugerMenu.css
@@ -13,7 +13,7 @@
     z-index: 10;
 }
 
-.hamburger-menu div {
+.hambuger-line{
     width: 100%;
     height: 3px;
     background-color: #333;
@@ -73,73 +73,4 @@
 
 .nav-menu ul li:hover {
     background-color: #f0f0f0;
-}
-
-/*ユーザー削除modal*/
-#user-delete-dialog {
-    border: none;
-    border-radius: 8px;
-    width: 400px;
-    height: auto; /* Adjust height to fit content */
-    z-index: 10;
-    padding: 20px;
-    box-sizing: border-box;
-    position: relative;
-}
-
-#user-delete-dialog::backdrop {
-    background: rgba(0,0,0,0.3);
-}
-
-#user-delete-dialog form div {
-    margin-bottom: 15px;
-}
-
-#user-delete-dialog label {
-    display: block;
-    margin-bottom: 5px;
-}
-
-#user-delete-dialog input {
-    width: calc(100% - 20px); /* Full width minus padding */
-    padding: 8px 10px;
-    border: 1px solid #ccc;
-    border-radius: 4px;
-    box-sizing: border-box;
-}
-
-#user-delete-dialog .input-success input {
-    border-color: lightgreen;
-}
-
-.error {
-    color: red;
-    margin-top: 10px;
-}
-
-#confirm-delete-btn {
-    margin-top: 20px;
-}
-
-#user-delete-modalBtn {
-    position: absolute;
-    top: 10px;
-    right: 10px;
-    background: none;
-    border: none;
-    font-size: 16px;
-    cursor: pointer; 
-}
-
-#confirm-delete-btn {
-    background: #007BFF;
-    color: white;
-    border: none;
-    border-radius: 4px;
-    padding: 8px 16px;
-    cursor: pointer;
-}
-
-#confirm-delete-btn:hover {
-    background: #0056b3;
 }
