+++ apps/frontend/my-app/src/Component/UserDelete.css
@@ -1,0 +1,68 @@
+/* UserDelete.css */
+#user-delete-dialog {
+    border: none;
+    border-radius: 8px;
+    width: 400px;
+    height: auto; /* コンテンツに合わせて高さを調整 */
+    z-index: 10;
+    padding: 20px;
+    box-sizing: border-box;
+    position: relative;
+}
+
+#user-delete-dialog::backdrop {
+    background: rgba(0,0,0,0.3);
+}
+
+#user-delete-dialog form div {
+    margin-bottom: 15px;
+}
+
+#user-delete-dialog label {
+    display: block;
+    margin-bottom: 5px;
+}
+
+#user-delete-dialog input {
+    width: calc(100% - 20px); /* パディングを引いた幅 */
+    padding: 8px 10px;
+    border: 1px solid #ccc;
+    border-radius: 4px;
+    box-sizing: border-box;
+}
+
+#user-delete-dialog .input-success input {
+    border-color: lightgreen;
+}
+
+.error {
+    color: red;
+    margin-top: 10px;
+}
+
+#user-delete-modalBtn, #confirm-delete-btn {
+    margin-top: 20px;
+}
+
+#user-delete-modalBtn {
+    position: absolute;
+    top: 10px;
+    right: 10px;
+    background: none;
+    border: none;
+    font-size: 16px;
+    cursor: pointer;
+}
+
+#confirm-delete-btn {
+    background: #007BFF;
+    color: white;
+    border: none;
+    border-radius: 4px;
+    padding: 8px 16px;
+    cursor: pointer;
+}
+
+#confirm-delete-btn:hover {
+    background: #0056b3;
+}
